Re-build project with postgreSQL
