+++ Mods/test/testtest.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<GameData>
+	<Units>
+		<Update>
+			<Where UnitType="null" />
+			<Set Combat ="0" Cost ="10" BaseMoves ="10" />
+		</Update>
+	</Units>
+</GameData>
