--- api/src/main/java/com/example/api/Api.java
@@ -1,16 +1,0 @@
-package com.example.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@SpringBootApplication
-public class Api {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Api.class, args);
-    }
-
-}
