--- api/src/main/java/com/example/api/Config/SecurityConfig.java
@@ -1,17 +1,0 @@
-package com.example.api.Config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-
-@EnableMethodSecurity
-@Configuration
-public class SecurityConfig {
-
-    @Bean
-    public WebSecurityCustomizer webSecurityCustomizer() {
-        return (web) -> web.ignoring().anyRequest();
-    }
-
-}
