--- api/src/main/java/com/example/api/Expection/ResourseNotFoundExpection.java
@@ -1,16 +1,0 @@
-package com.example.api.Expection;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.io.Serial;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourseNotFoundExpection extends RuntimeException{
-    @Serial
-    private static final long serialVersionUID = 1L;
-    public ResourseNotFoundExpection(String message){
-        super(message);
-    }
-
-}
