--- api/src/main/java/com/example/api/Repository/mod_unitsRepository.java
@@ -1,9 +1,0 @@
-package com.example.api.Repository;
-
-import com.example.api.Model.Mods_units;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface mod_unitsRepository extends JpaRepository<Mods_units, Long> {
-}
