--- api/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/db
-spring.datasource.username=root
-spring.datasource.password=
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.database-platform = org.hibernate.dialect.MySQL8Dialect
-spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto = update
-spring.application.name = spring_boot_example
-server.servlet.context-path=/api/v1
-spring.jpa.defer-datasource-initialization = true
-spring.sql.init.mode = always
-
