--- api/src/test/java/com/example/api/ApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.api;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
