+++ src/main/java/com/civmodapi/Config/SecurityConfig.java
@@ -1,0 +1,17 @@
+package com.civmodapi.Config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+
+@EnableMethodSecurity
+@Configuration
+public class SecurityConfig {
+
+    @Bean
+    public WebSecurityCustomizer webSecurityCustomizer() {
+        return (web) -> web.ignoring().anyRequest();
+    }
+
+}
