+++ src/main/java/com/civmodapi/Img/Slinger_(Civ6).png
