+++ src/main/java/com/civmodapi/Img/Warrior_(Civ6).png
