+++ src/main/resources/flyway.conf
@@ -1,0 +1,3 @@
+flyway.user=root
+flyway.password=123
+flyway.schemas=jdbc:postgresql://localhost:5432/dbCiv
