+++ src/test/java/com/civmodapi/CivModApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.civmodapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CivModApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
